requirements.txt
pytest==8.3.5
python-dotenv==1.0.1
pytest-check==2.5.0
openpyxl==3.1.5
selenium==4.29.0
webdriver_manager==4.0.2
Pillow==11.1.0
